% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconData-package.R
\docType{data}
\name{bowl18}
\alias{bowl18}
\title{Bowling et al. (2018)}
\description{
This dataset comes from the supplementary materials of
\insertCite{Bowling2018;textual}{inconData}.
}
\details{
The dataset is represented as a \link[tibble:tibble-package]{tibble}
with the following columns:
\itemize{
\item \code{id}: Unique ID for each chord.
\item \code{name}: Chord name (only present for some chords).
\item \code{chord_size}: Number of chord tones.
\item \code{rating_mean}:
Mean consonance rating. Ratings were on a four-point scale,
where high ratings indicated high consonance.
\item \code{rating_sd}: Standard deviation of consonance ratings.
\item \code{rating_se}: Standard error of consonance ratings.
\item \code{fr_chord}: Vector of chord frequencies.
\item \code{pi_chord}: Vector of chord pitches, on MIDI note scale.
\item \code{pi_chord_int}: An integer approximation of \code{pi_chord}.
\item \code{pi_chord_type_int}:
\code{pi_chord_int} but expressed relative to the chord's bass note.
}
}
\references{
\insertAllCited
}
\keyword{data}
