% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incon-datasets.R
\docType{data}
\name{incon_datasets}
\alias{incon_datasets}
\title{Simultaneous consonance datasets}
\usage{
incon_datasets()
}
\description{
This \code{\link[tibble]{tibble}} summarises the
consonance perception datasets available in the \code{inconData} package.
\itemize{
\item \code{label}: Label for the dataset.
\item \code{description}: Description of the dataset.
}
}
\keyword{data}
