% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconData-package.R
\docType{package}
\name{inconData-package}
\alias{inconData}
\alias{inconData-package}
\title{inconData: Datasets for Simultaneous Consonance Perception}
\description{
This package compiles several empirical datasets for experiments on simultaneous consonance perception.
}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
