% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl12.R
\docType{data}
\name{jl12}
\alias{jl12}
\alias{jl12a}
\alias{jl12b}
\title{Johnson-Laird et al. (2012)}
\description{
These datasets were transcribed from
\insertCite{Johnson-Laird2012;textual}{inconData}.
}
\details{
Two datasets are provided:
\itemize{
\item \code{jl12a}: corresponds to Figure 2 of
\insertCite{Johnson-Laird2012;textual}{inconData}.
\item \code{jl12b}: corresponds to Figure 3 of
\insertCite{Johnson-Laird2012;textual}{inconData}.
}

Each dataset is represented as a \link[tibble:tibble-package]{tibble}
with the following columns:
\itemize{
\item \code{id}: Unique chord identifer.
\item \code{perceptual_dissonance}: Mean perceptual dissonance rating.
\item \code{roughness_model}:
These are estimates of roughness made by
\insertCite{Johnson-Laird2012;textual}{inconData}
using the computational model of
\insertCite{Hutchinson1978;textual}{inconData}.
There seem to be some mistakes in this data;
an alternative is to use the model as implemented in
the \code{\link[incon]{incon}} package.
\item \code{tonal_model}:
These are consonance ranks from the tonal consonance model of
\insertCite{Johnson-Laird2012;textual}{inconData}.
Lower values correspond to greater consonance.
Again, there seem to be some mistakes in this data;
an alternative is to use the model as implemented in
the \code{\link[incon]{incon}} package.
\item \code{pi_chord}:
Vector of chord pitches, expressed as MIDI note numbers.
Uses the \code{\link[hrep]{pi_chord}} representation from the
\code{hrep} package.
}
}
\note{
The table uses representations from the \code{hrep} package.
Loading the \code{hrep} package
improves the way that chord objects are displayed.
}
\references{
\insertAllCited
}
\keyword{data}
