% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lah16.R
\docType{data}
\name{lah16}
\alias{lah16}
\title{Lahdelma & Eerola (2016)}
\description{
This dataset is reproduced from the supplementary materials of
\insertCite{Lahdelma2016;textual}{inconData}.
}
\details{
The dataset is represented as a \link[tibble:tibble-package]{tibble}
with the following columns:
\itemize{
\item \code{label}: Chord label.
\item \code{num_ratings}: Number of times that the chord was rated.
\item \code{consonance_rating_mean}: Mean consonance rating.
Higher values indicate higher consonance.
\item \code{consonance_rating_sd}: Standard deviation of consonance ratings.
\item \code{consonance_rating_se}: Standard error of the
mean consonance rating, computed under the central limit theorem.
\item \code{pi_chord_type}
Vector of chord pitches, each expressed relative to the bass pitch.
Uses the \code{\link[hrep]{pi_chord_type}} representation from the
\code{hrep} package.
}
}
\note{
It seems that the original authors did not store transposition information
along with this dataset.

The table uses representations from the \code{hrep} package.
Loading the \code{hrep} package
improves the way that chord objects are displayed.
}
\references{
\insertAllCited
}
\keyword{data}
